Update vim-minimal in the CentOS 6 image.
The CentOS 6 (x86_64) image is affected by the problem described at
https://bugzilla.redhat.com/show_bug.cgi?id=1329015. Which is apparently
an old notabug present in RedHat/Fedora distributions, manifesting
itself as a conflict on attempt to install a newer vim-common on a
system where an outdated version of vim-minimal is installed.

The common resolution is to update vim-minimal before installing
vim-common. Which is easier to do in the image itself.