Remove Debian Jessie backports repositories
It was removed from the official mirror, because backports is not a part
of the LTS architecture. See [1] and [2] ('Deprecation of LTS support
for backports') for more information.

The reason to update the image is that doing `apt-get update` from the
image (say, from a prebuild script) now leads to the following errors:

 | Err http://ftp.debian.org jessie-backports/main Sources
 |   404  Not Found
 | Err http://ftp.debian.org jessie-backports/main amd64 Packages
 |   404  Not Found

[1]: https://lists.debian.org/debian-devel-announce/2019/03/msg00006.html
[2]: https://backports.debian.org/