Update Fedora Rawhide Dockerfile
Now it is more or less same as Fedora 31 Dockerfile.

This commit allows to successfully build the image. Before it fails,
because there is no more pygpgme package in Fedora Rawhide repositories.

Commented backports repository enabling, because the packagecloud.io
script detects the operating system as fedora/32 and reports that does
not support it.

The problem with Python 3 was resolved a long time ago, so the
workaround for it is not needed anymore.