Add Dockerfile for Debian 11 (bullseye)
Removed installation of dh-systemd, as this package is not a part of
Debian 11 anymore. It's functionality included into debhelper, which is
installed anyway.

@Totktonada: Dropped --force-yes, because it is deprecated in favor of
--allow-* parameters and because everything works without any of those
parameters. I didn't find a reason why it was added in the first place.
I guess it was attempt to make apt-get fully non-interactive. Let's come
back here if we'll meet any real problem.

Co-authored-by: Alexander Turenko <alexander.turenko@tarantool.org>