Update repositories and installed packages
Found that on old OS with old metadata in cache a lot of connections
hangs on the first packages update. On second packages update command
fails with errors because of hanged connections, like [1]:

  237   connect(11, {sa_family=AF_INET6, sin6_port=htons(443), sin6_flowinfo=htonl(0), inet_pton(AF_INET6, "2605:bc80:3010:600:dead:beef:cafe:fed9", &sin6_addr), sin6_scope_id=0}, 28 <unfinished ...>
  236   rt_sigaction(SIGPIPE, NULL,  <unfinished ...>
  237   <... connect resumed>)            = -1 ENETUNREACH (Network is unreachable)

To avoid of it on Centos/Fedora images repositories and packages update
call added. Also on Fedora 29/30 images repositories configurations
cleaned from deprecated option 'failovermethod'.

Closes tarantool/tarantool-qa#60

[1]: https://github.com/tarantool/tarantool-qa/issues/60#issuecomment-789596820