Add RedOS 7.3
It is RPM based Linux distribution.

The RedOS image is built using the multistage build process. CentOS 7 is
used as the builder container: it calls `yum --installroot=<...>
install` to install packages into the root of the target RedOS image
(the idea is similar to chroot).

This way we're able to create RedOS image without extra scripting around
`docker build` command. It seems, it is not possible if we'll use the
official RedOS docker image archive (because it is not in a docker
registry, just a file).

https://redos.red-soft.ru

Co-authored-by: Alexander Turenko <alexander.turenko@tarantool.org>